Split the cross model facade into a controller facade and model facade
