+++ api/facadeversions.go
@@ -18,6 +18,7 @@ var facadeVersions = map[string]int{
 	"AllWatcher":                   1,
 	"Annotations":                  2,
 	"Application":                  4,
+	"ApplicationOffers":            1,
 	"ApplicationScaler":            1,
 	"Backups":                      1,
 	"Block":                        2,
@@ -28,7 +29,6 @@ var facadeVersions = map[string]int{
 	"Client":                       1,
 	"Cloud":                        1,
 	"Controller":                   3,
-	"CrossModelRelations":          1,
 	"Deployer":                     1,
 	"DiscoverSpaces":               2,
 	"DiskManager":                  2,
@@ -69,6 +69,7 @@ var facadeVersions = map[string]int{
 	"ProxyUpdater":                 1,
 	"Reboot":                       2,
 	"RelationUnitsWatcher":         1,
+	"RemoteEndpoints":              1,
 	"RemoteFirewaller":             1,
 	"RemoteRelations":              1,
 	"Resources":                    1,
