+++ apiserver/applicationoffers/applicationoffers_test.go
@@ -1,0 +1,221 @@
+// Copyright 2015 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package applicationoffers_test
+
+import (
+	"fmt"
+
+	"github.com/juju/errors"
+	jc "github.com/juju/testing/checkers"
+	gc "gopkg.in/check.v1"
+	"gopkg.in/juju/charm.v6-unstable"
+	"gopkg.in/juju/names.v2"
+
+	"github.com/juju/juju/apiserver/applicationoffers"
+	"github.com/juju/juju/apiserver/common"
+	"github.com/juju/juju/apiserver/common/crossmodelcommon"
+	"github.com/juju/juju/apiserver/params"
+	jujucrossmodel "github.com/juju/juju/core/crossmodel"
+)
+
+type applicationOffersSuite struct {
+	baseSuite
+	api *applicationoffers.OffersAPI
+}
+
+var _ = gc.Suite(&applicationOffersSuite{})
+
+func (s *applicationOffersSuite) SetUpTest(c *gc.C) {
+	s.baseSuite.SetUpTest(c)
+	s.applicationOffers = &mockApplicationOffers{}
+	getApplicationOffers := func(interface{}) jujucrossmodel.ApplicationOffers {
+		return s.applicationOffers
+	}
+
+	var err error
+	s.api, err = applicationoffers.CreateOffersAPI(
+		getApplicationOffers, s.mockState, s.mockStatePool, s.authorizer,
+	)
+	c.Assert(err, jc.ErrorIsNil)
+}
+
+func (s *applicationOffersSuite) assertOffer(c *gc.C, expectedErr error) {
+	applicationName := "test"
+	s.addApplication(c, applicationName)
+	one := params.AddApplicationOffer{
+		OfferName:       "offer-test",
+		ApplicationName: applicationName,
+		Endpoints:       map[string]string{"db": "db"},
+	}
+	all := params.AddApplicationOffers{Offers: []params.AddApplicationOffer{one}}
+	s.applicationOffers.addOffer = func(offer jujucrossmodel.AddApplicationOfferArgs) (*jujucrossmodel.ApplicationOffer, error) {
+		c.Assert(offer.OfferName, gc.Equals, one.OfferName)
+		c.Assert(offer.ApplicationName, gc.Equals, one.ApplicationName)
+		c.Assert(offer.ApplicationDescription, gc.Equals, "A pretty popular blog engine")
+		return &jujucrossmodel.ApplicationOffer{}, nil
+	}
+	charm := &mockCharm{meta: &charm.Meta{Description: "A pretty popular blog engine"}}
+	s.mockState.applications = map[string]crossmodelcommon.Application{
+		applicationName: &mockApplication{charm: charm},
+	}
+
+	errs, err := s.api.Offer(all)
+	if expectedErr != nil {
+		c.Assert(errors.Cause(err), gc.ErrorMatches, expectedErr.Error())
+		return
+	}
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(errs.Results, gc.HasLen, len(all.Offers))
+	c.Assert(errs.Results[0].Error, gc.IsNil)
+	s.applicationOffers.CheckCallNames(c, addOffersBackendCall)
+}
+
+func (s *applicationOffersSuite) TestOffer(c *gc.C) {
+	s.authorizer.Tag = names.NewUserTag("admin")
+	s.assertOffer(c, nil)
+}
+
+func (s *applicationOffersSuite) TestOfferPermission(c *gc.C) {
+	s.assertOffer(c, common.ErrPerm)
+}
+
+func (s *applicationOffersSuite) TestOfferSomeFail(c *gc.C) {
+	s.authorizer.Tag = names.NewUserTag("admin")
+	s.addApplication(c, "one")
+	s.addApplication(c, "two")
+	s.addApplication(c, "paramsfail")
+	one := params.AddApplicationOffer{
+		OfferName:       "offer-one",
+		ApplicationName: "one",
+		Endpoints:       map[string]string{"db": "db"},
+	}
+	bad := params.AddApplicationOffer{
+		OfferName:       "offer-bad",
+		ApplicationName: "notthere",
+		Endpoints:       map[string]string{"db": "db"},
+	}
+	bad2 := params.AddApplicationOffer{
+		OfferName:       "offer-bad",
+		ApplicationName: "paramsfail",
+		Endpoints:       map[string]string{"db": "db"},
+	}
+	two := params.AddApplicationOffer{
+		OfferName:       "offer-two",
+		ApplicationName: "two",
+		Endpoints:       map[string]string{"db": "db"},
+	}
+	all := params.AddApplicationOffers{Offers: []params.AddApplicationOffer{one, bad, bad2, two}}
+	s.applicationOffers.addOffer = func(offer jujucrossmodel.AddApplicationOfferArgs) (*jujucrossmodel.ApplicationOffer, error) {
+		if offer.ApplicationName == "paramsfail" {
+			return nil, errors.New("params fail")
+		}
+		return &jujucrossmodel.ApplicationOffer{}, nil
+	}
+	charm := &mockCharm{meta: &charm.Meta{Description: "A pretty popular blog engine"}}
+	s.mockState.applications = map[string]crossmodelcommon.Application{
+		"one":        &mockApplication{charm: charm},
+		"two":        &mockApplication{charm: charm},
+		"paramsfail": &mockApplication{charm: charm},
+	}
+
+	errs, err := s.api.Offer(all)
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(errs.Results, gc.HasLen, len(all.Offers))
+	c.Assert(errs.Results[0].Error, gc.IsNil)
+	c.Assert(errs.Results[3].Error, gc.IsNil)
+	c.Assert(errs.Results[1].Error, gc.ErrorMatches, `getting offered application notthere: application "notthere" not found`)
+	c.Assert(errs.Results[2].Error, gc.ErrorMatches, `params fail`)
+	s.applicationOffers.CheckCallNames(c, addOffersBackendCall, addOffersBackendCall, addOffersBackendCall)
+}
+
+func (s *applicationOffersSuite) TestOfferError(c *gc.C) {
+	s.authorizer.Tag = names.NewUserTag("admin")
+	applicationName := "test"
+	s.addApplication(c, applicationName)
+	one := params.AddApplicationOffer{
+		OfferName:       "offer-test",
+		ApplicationName: applicationName,
+		Endpoints:       map[string]string{"db": "db"},
+	}
+	all := params.AddApplicationOffers{Offers: []params.AddApplicationOffer{one}}
+
+	msg := "fail"
+
+	s.applicationOffers.addOffer = func(offer jujucrossmodel.AddApplicationOfferArgs) (*jujucrossmodel.ApplicationOffer, error) {
+		return nil, errors.New(msg)
+	}
+	charm := &mockCharm{meta: &charm.Meta{Description: "A pretty popular blog engine"}}
+	s.mockState.applications = map[string]crossmodelcommon.Application{
+		applicationName: &mockApplication{charm: charm},
+	}
+
+	errs, err := s.api.Offer(all)
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(errs.Results, gc.HasLen, len(all.Offers))
+	c.Assert(errs.Results[0].Error, gc.ErrorMatches, fmt.Sprintf(".*%v.*", msg))
+	s.applicationOffers.CheckCallNames(c, addOffersBackendCall)
+}
+
+func (s *applicationOffersSuite) assertList(c *gc.C, expectedErr error) {
+	s.setupOffers(c, "test")
+	filter := params.OfferFilters{
+		Filters: []params.OfferFilter{
+			{
+				OfferName:       "hosted-db2",
+				ApplicationName: "test",
+			},
+		},
+	}
+	found, err := s.api.ListApplicationOffers(filter)
+	if expectedErr != nil {
+		c.Assert(errors.Cause(err), gc.ErrorMatches, expectedErr.Error())
+		return
+	}
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(found, jc.DeepEquals, params.ListApplicationOffersResults{
+		[]params.ApplicationOfferDetails{
+			{
+				ApplicationOffer: params.ApplicationOffer{
+					ApplicationDescription: "description",
+					OfferName:              "hosted-db2",
+					OfferURL:               "fred/prod.hosted-db2",
+					Endpoints:              []params.RemoteEndpoint{{Name: "db"}},
+				},
+				CharmName:      "db2",
+				ConnectedCount: 5,
+			},
+		},
+	})
+	s.applicationOffers.CheckCallNames(c, listOffersBackendCall)
+}
+
+func (s *applicationOffersSuite) TestList(c *gc.C) {
+	s.authorizer.Tag = names.NewUserTag("admin")
+	s.assertList(c, nil)
+}
+
+func (s *applicationOffersSuite) TestListPermission(c *gc.C) {
+	s.assertList(c, common.ErrPerm)
+}
+
+func (s *applicationOffersSuite) TestListError(c *gc.C) {
+	s.authorizer.Tag = names.NewUserTag("admin")
+	filter := params.OfferFilters{
+		Filters: []params.OfferFilter{
+			{
+				OfferName:       "hosted-db2",
+				ApplicationName: "test",
+			},
+		},
+	}
+	msg := "fail"
+
+	s.applicationOffers.listOffers = func(filters ...jujucrossmodel.ApplicationOfferFilter) ([]jujucrossmodel.ApplicationOffer, error) {
+		return nil, errors.New(msg)
+	}
+
+	_, err := s.api.ListApplicationOffers(filter)
+	c.Assert(err, gc.ErrorMatches, fmt.Sprintf(".*%v.*", msg))
+	s.applicationOffers.CheckCallNames(c, listOffersBackendCall)
+}
