+++ apiserver/applicationoffers/export_test.go
@@ -1,8 +1,8 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package crossmodel
+package applicationoffers
 
 var (
-	CreateAPI = createAPI
+	CreateOffersAPI = createOffersAPI
 )
