--- apiserver/crossmodel/package_test.go
@@ -1,14 +1,0 @@
-// Copyright 2015 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package crossmodel_test
-
-import (
-	stdtesting "testing"
-
-	"github.com/juju/juju/testing"
-)
-
-func TestAll(t *stdtesting.T) {
-	testing.MgoTestPackage(t)
-}
