+++ apiserver/remoteendpoints/remoteendpoints_test.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package crossmodel_test
+package remoteendpoints_test
 
 import (
 	"fmt"
@@ -13,135 +13,34 @@ import (
 	"gopkg.in/juju/names.v2"
 
 	"github.com/juju/juju/apiserver/common"
-	"github.com/juju/juju/apiserver/crossmodel"
+	"github.com/juju/juju/apiserver/common/crossmodelcommon"
 	"github.com/juju/juju/apiserver/params"
+	"github.com/juju/juju/apiserver/remoteendpoints"
 	jujucrossmodel "github.com/juju/juju/core/crossmodel"
 )
 
-type crossmodelSuite struct {
-	baseCrossmodelSuite
+type remoteEndpointsSuite struct {
+	baseSuite
+	api *remoteendpoints.EndpointsAPI
 }
 
-var _ = gc.Suite(&crossmodelSuite{})
+var _ = gc.Suite(&remoteEndpointsSuite{})
 
-func (s *crossmodelSuite) assertOffer(c *gc.C, expectedErr error) {
-	applicationName := "test"
-	s.addApplication(c, applicationName)
-	one := params.AddApplicationOffer{
-		OfferName:       "offer-test",
-		ApplicationName: applicationName,
-		Endpoints:       map[string]string{"db": "db"},
-	}
-	all := params.AddApplicationOffers{Offers: []params.AddApplicationOffer{one}}
-	s.applicationOffers.addOffer = func(offer jujucrossmodel.AddApplicationOfferArgs) (*jujucrossmodel.ApplicationOffer, error) {
-		c.Assert(offer.OfferName, gc.Equals, one.OfferName)
-		c.Assert(offer.ApplicationName, gc.Equals, one.ApplicationName)
-		c.Assert(offer.ApplicationDescription, gc.Equals, "A pretty popular blog engine")
-		return &jujucrossmodel.ApplicationOffer{}, nil
-	}
-	charm := &mockCharm{meta: &charm.Meta{Description: "A pretty popular blog engine"}}
-	s.mockState.applications = map[string]crossmodel.Application{
-		applicationName: &mockApplication{charm: charm},
+func (s *remoteEndpointsSuite) SetUpTest(c *gc.C) {
+	s.baseSuite.SetUpTest(c)
+	s.applicationOffers = &mockApplicationOffers{}
+	getApplicationOffers := func(interface{}) jujucrossmodel.ApplicationOffers {
+		return s.applicationOffers
 	}
 
-	errs, err := s.api.Offer(all)
-	if expectedErr != nil {
-		c.Assert(errors.Cause(err), gc.ErrorMatches, expectedErr.Error())
-		return
-	}
+	var err error
+	s.api, err = remoteendpoints.CreateEndpointsAPI(
+		getApplicationOffers, s.mockState, s.mockStatePool, s.authorizer,
+	)
 	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(errs.Results, gc.HasLen, len(all.Offers))
-	c.Assert(errs.Results[0].Error, gc.IsNil)
-	s.applicationOffers.CheckCallNames(c, addOffersBackendCall)
 }
 
-func (s *crossmodelSuite) TestOffer(c *gc.C) {
-	s.authorizer.Tag = names.NewUserTag("admin")
-	s.assertOffer(c, nil)
-}
-
-func (s *crossmodelSuite) TestOfferPermission(c *gc.C) {
-	s.assertOffer(c, common.ErrPerm)
-}
-
-func (s *crossmodelSuite) TestOfferSomeFail(c *gc.C) {
-	s.authorizer.Tag = names.NewUserTag("admin")
-	s.addApplication(c, "one")
-	s.addApplication(c, "two")
-	s.addApplication(c, "paramsfail")
-	one := params.AddApplicationOffer{
-		OfferName:       "offer-one",
-		ApplicationName: "one",
-		Endpoints:       map[string]string{"db": "db"},
-	}
-	bad := params.AddApplicationOffer{
-		OfferName:       "offer-bad",
-		ApplicationName: "notthere",
-		Endpoints:       map[string]string{"db": "db"},
-	}
-	bad2 := params.AddApplicationOffer{
-		OfferName:       "offer-bad",
-		ApplicationName: "paramsfail",
-		Endpoints:       map[string]string{"db": "db"},
-	}
-	two := params.AddApplicationOffer{
-		OfferName:       "offer-two",
-		ApplicationName: "two",
-		Endpoints:       map[string]string{"db": "db"},
-	}
-	all := params.AddApplicationOffers{Offers: []params.AddApplicationOffer{one, bad, bad2, two}}
-	s.applicationOffers.addOffer = func(offer jujucrossmodel.AddApplicationOfferArgs) (*jujucrossmodel.ApplicationOffer, error) {
-		if offer.ApplicationName == "paramsfail" {
-			return nil, errors.New("params fail")
-		}
-		return &jujucrossmodel.ApplicationOffer{}, nil
-	}
-	charm := &mockCharm{meta: &charm.Meta{Description: "A pretty popular blog engine"}}
-	s.mockState.applications = map[string]crossmodel.Application{
-		"one":        &mockApplication{charm: charm},
-		"two":        &mockApplication{charm: charm},
-		"paramsfail": &mockApplication{charm: charm},
-	}
-
-	errs, err := s.api.Offer(all)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(errs.Results, gc.HasLen, len(all.Offers))
-	c.Assert(errs.Results[0].Error, gc.IsNil)
-	c.Assert(errs.Results[3].Error, gc.IsNil)
-	c.Assert(errs.Results[1].Error, gc.ErrorMatches, `getting offered application notthere: application "notthere" not found`)
-	c.Assert(errs.Results[2].Error, gc.ErrorMatches, `params fail`)
-	s.applicationOffers.CheckCallNames(c, addOffersBackendCall, addOffersBackendCall, addOffersBackendCall)
-}
-
-func (s *crossmodelSuite) TestOfferError(c *gc.C) {
-	s.authorizer.Tag = names.NewUserTag("admin")
-	applicationName := "test"
-	s.addApplication(c, applicationName)
-	one := params.AddApplicationOffer{
-		OfferName:       "offer-test",
-		ApplicationName: applicationName,
-		Endpoints:       map[string]string{"db": "db"},
-	}
-	all := params.AddApplicationOffers{Offers: []params.AddApplicationOffer{one}}
-
-	msg := "fail"
-
-	s.applicationOffers.addOffer = func(offer jujucrossmodel.AddApplicationOfferArgs) (*jujucrossmodel.ApplicationOffer, error) {
-		return nil, errors.New(msg)
-	}
-	charm := &mockCharm{meta: &charm.Meta{Description: "A pretty popular blog engine"}}
-	s.mockState.applications = map[string]crossmodel.Application{
-		applicationName: &mockApplication{charm: charm},
-	}
-
-	errs, err := s.api.Offer(all)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(errs.Results, gc.HasLen, len(all.Offers))
-	c.Assert(errs.Results[0].Error, gc.ErrorMatches, fmt.Sprintf(".*%v.*", msg))
-	s.applicationOffers.CheckCallNames(c, addOffersBackendCall)
-}
-
-func (s *crossmodelSuite) assertShow(c *gc.C, expectedErr error) {
+func (s *remoteEndpointsSuite) assertShow(c *gc.C, expectedErr error) {
 	applicationName := "test"
 	offerName := "hosted-test"
 	url := "fred/prod.hosted-db2"
@@ -158,11 +57,11 @@ func (s *crossmodelSuite) assertShow(c *gc.C, expectedErr error) {
 		return []jujucrossmodel.ApplicationOffer{anOffer}, nil
 	}
 	ch := &mockCharm{meta: &charm.Meta{Description: "A pretty popular blog engine"}}
-	s.mockState.applications = map[string]crossmodel.Application{
+	s.mockState.applications = map[string]crossmodelcommon.Application{
 		applicationName: &mockApplication{charm: ch, curl: charm.MustParseURL("db2-2")},
 	}
 	s.mockState.model = &mockModel{uuid: "uuid", name: "prod", owner: "fred"}
-	s.mockState.usermodels = []crossmodel.UserModel{
+	s.mockState.usermodels = []crossmodelcommon.UserModel{
 		&mockUserModel{model: s.mockState.model},
 	}
 	s.mockState.connStatus = &mockConnectionStatus{count: 5}
@@ -185,16 +84,16 @@ func (s *crossmodelSuite) assertShow(c *gc.C, expectedErr error) {
 	s.applicationOffers.CheckCallNames(c, listOffersBackendCall)
 }
 
-func (s *crossmodelSuite) TestShow(c *gc.C) {
+func (s *remoteEndpointsSuite) TestShow(c *gc.C) {
 	s.assertShow(c, nil)
 }
 
-func (s *crossmodelSuite) TestShowPermission(c *gc.C) {
+func (s *remoteEndpointsSuite) TestShowPermission(c *gc.C) {
 	s.authorizer.Tag = names.NewUserTag("someone")
 	s.assertShow(c, common.ErrPerm)
 }
 
-func (s *crossmodelSuite) TestShowError(c *gc.C) {
+func (s *remoteEndpointsSuite) TestShowError(c *gc.C) {
 	url := "fred/prod.hosted-db2"
 	filter := params.ApplicationURLs{[]string{url}}
 	msg := "fail"
@@ -203,7 +102,7 @@ func (s *crossmodelSuite) TestShowError(c *gc.C) {
 		return nil, errors.New(msg)
 	}
 	s.mockState.model = &mockModel{uuid: "uuid", name: "prod", owner: "fred"}
-	s.mockState.usermodels = []crossmodel.UserModel{
+	s.mockState.usermodels = []crossmodelcommon.UserModel{
 		&mockUserModel{model: s.mockState.model},
 	}
 
@@ -214,7 +113,7 @@ func (s *crossmodelSuite) TestShowError(c *gc.C) {
 	s.applicationOffers.CheckCallNames(c, listOffersBackendCall)
 }
 
-func (s *crossmodelSuite) TestShowNotFound(c *gc.C) {
+func (s *remoteEndpointsSuite) TestShowNotFound(c *gc.C) {
 	urls := []string{"fred/prod.hosted-db2"}
 	filter := params.ApplicationURLs{urls}
 
@@ -222,7 +121,7 @@ func (s *crossmodelSuite) TestShowNotFound(c *gc.C) {
 		return nil, nil
 	}
 	s.mockState.model = &mockModel{uuid: "uuid", name: "prod", owner: "fred"}
-	s.mockState.usermodels = []crossmodel.UserModel{
+	s.mockState.usermodels = []crossmodelcommon.UserModel{
 		&mockUserModel{model: s.mockState.model},
 	}
 
@@ -233,7 +132,7 @@ func (s *crossmodelSuite) TestShowNotFound(c *gc.C) {
 	s.applicationOffers.CheckCallNames(c, listOffersBackendCall)
 }
 
-func (s *crossmodelSuite) TestShowErrorMsgMultipleURLs(c *gc.C) {
+func (s *remoteEndpointsSuite) TestShowErrorMsgMultipleURLs(c *gc.C) {
 	urls := []string{"fred/prod.hosted-mysql", "fred/test.hosted-db2"}
 	filter := params.ApplicationURLs{urls}
 
@@ -241,7 +140,7 @@ func (s *crossmodelSuite) TestShowErrorMsgMultipleURLs(c *gc.C) {
 		return nil, nil
 	}
 	s.mockState.model = &mockModel{uuid: "uuid", name: "prod", owner: "fred"}
-	s.mockState.usermodels = []crossmodel.UserModel{
+	s.mockState.usermodels = []crossmodelcommon.UserModel{
 		&mockUserModel{model: s.mockState.model},
 		&mockUserModel{model: &mockModel{uuid: "uuid2", name: "test", owner: "fred"}},
 	}
@@ -256,7 +155,7 @@ func (s *crossmodelSuite) TestShowErrorMsgMultipleURLs(c *gc.C) {
 	s.applicationOffers.CheckCallNames(c, listOffersBackendCall, listOffersBackendCall)
 }
 
-func (s *crossmodelSuite) TestShowFoundMultiple(c *gc.C) {
+func (s *remoteEndpointsSuite) TestShowFoundMultiple(c *gc.C) {
 	name := "test"
 	url := "fred/prod.hosted-sql"
 	anOffer := jujucrossmodel.ApplicationOffer{
@@ -285,18 +184,18 @@ func (s *crossmodelSuite) TestShowFoundMultiple(c *gc.C) {
 		return []jujucrossmodel.ApplicationOffer{anOffer2}, nil
 	}
 	ch := &mockCharm{meta: &charm.Meta{Description: "A pretty popular blog engine"}}
-	s.mockState.applications = map[string]crossmodel.Application{
+	s.mockState.applications = map[string]crossmodelcommon.Application{
 		"test": &mockApplication{charm: ch, curl: charm.MustParseURL("db2-2")},
 	}
 	s.mockState.model = &mockModel{uuid: "uuid", name: "prod", owner: "fred"}
-	s.mockState.usermodels = []crossmodel.UserModel{
+	s.mockState.usermodels = []crossmodelcommon.UserModel{
 		&mockUserModel{model: s.mockState.model},
 		&mockUserModel{model: &mockModel{uuid: "uuid2", name: "test", owner: "mary"}},
 	}
 	s.mockState.connStatus = &mockConnectionStatus{count: 5}
 
 	anotherState := &mockState{modelUUID: "uuid2"}
-	anotherState.applications = map[string]crossmodel.Application{
+	anotherState.applications = map[string]crossmodelcommon.Application{
 		"testAgain": &mockApplication{charm: ch, curl: charm.MustParseURL("mysql-2")},
 	}
 	anotherState.connStatus = &mockConnectionStatus{count: 5}
@@ -324,37 +223,7 @@ func (s *crossmodelSuite) TestShowFoundMultiple(c *gc.C) {
 	s.applicationOffers.CheckCallNames(c, listOffersBackendCall, listOffersBackendCall)
 }
 
-func (s *crossmodelSuite) setupOffers(c *gc.C, filterAppName string) {
-	applicationName := "test"
-	offerName := "hosted-db2"
-
-	anOffer := jujucrossmodel.ApplicationOffer{
-		OfferName:              offerName,
-		ApplicationName:        applicationName,
-		ApplicationDescription: "description",
-		Endpoints:              map[string]charm.Relation{"db": {Name: "db2"}},
-	}
-
-	s.applicationOffers.listOffers = func(filters ...jujucrossmodel.ApplicationOfferFilter) ([]jujucrossmodel.ApplicationOffer, error) {
-		c.Assert(filters, gc.HasLen, 1)
-		c.Assert(filters[0], jc.DeepEquals, jujucrossmodel.ApplicationOfferFilter{
-			OfferName:       offerName,
-			ApplicationName: filterAppName,
-		})
-		return []jujucrossmodel.ApplicationOffer{anOffer}, nil
-	}
-	ch := &mockCharm{meta: &charm.Meta{Description: "A pretty popular blog engine"}}
-	s.mockState.applications = map[string]crossmodel.Application{
-		"test": &mockApplication{charm: ch, curl: charm.MustParseURL("db2-2")},
-	}
-	s.mockState.model = &mockModel{uuid: "uuid", name: "prod", owner: "fred"}
-	s.mockState.usermodels = []crossmodel.UserModel{
-		&mockUserModel{model: s.mockState.model},
-	}
-	s.mockState.connStatus = &mockConnectionStatus{count: 5}
-}
-
-func (s *crossmodelSuite) assertFind(c *gc.C, expectedErr error) {
+func (s *remoteEndpointsSuite) assertFind(c *gc.C, expectedErr error) {
 	s.setupOffers(c, "")
 	filter := params.OfferFilters{
 		Filters: []params.OfferFilter{
@@ -380,16 +249,16 @@ func (s *crossmodelSuite) assertFind(c *gc.C, expectedErr error) {
 	s.applicationOffers.CheckCallNames(c, listOffersBackendCall)
 }
 
-func (s *crossmodelSuite) TestFind(c *gc.C) {
+func (s *remoteEndpointsSuite) TestFind(c *gc.C) {
 	s.assertFind(c, nil)
 }
 
-func (s *crossmodelSuite) TestFindPermission(c *gc.C) {
+func (s *remoteEndpointsSuite) TestFindPermission(c *gc.C) {
 	s.authorizer.Tag = names.NewUserTag("someone")
 	s.assertFind(c, common.ErrPerm)
 }
 
-func (s *crossmodelSuite) TestFindMultiModel(c *gc.C) {
+func (s *remoteEndpointsSuite) TestFindMultiModel(c *gc.C) {
 	db2Offer := jujucrossmodel.ApplicationOffer{
 		OfferName:              "hosted-db2",
 		ApplicationName:        "db2",
@@ -424,7 +293,7 @@ func (s *crossmodelSuite) TestFindMultiModel(c *gc.C) {
 		return result, nil
 	}
 	ch := &mockCharm{meta: &charm.Meta{Description: "A pretty popular blog engine"}}
-	s.mockState.applications = map[string]crossmodel.Application{
+	s.mockState.applications = map[string]crossmodelcommon.Application{
 		"db2": &mockApplication{charm: ch, curl: charm.MustParseURL("db2-2")},
 	}
 	s.mockState.model = &mockModel{uuid: "uuid", name: "prod", owner: "fred"}
@@ -432,17 +301,17 @@ func (s *crossmodelSuite) TestFindMultiModel(c *gc.C) {
 
 	anotherState := &mockState{modelUUID: "uuid2"}
 	s.mockStatePool.st["uuid2"] = anotherState
-	anotherState.applications = map[string]crossmodel.Application{
+	anotherState.applications = map[string]crossmodelcommon.Application{
 		"mysql":      &mockApplication{charm: ch, curl: charm.MustParseURL("mysql-2")},
 		"postgresql": &mockApplication{charm: ch, curl: charm.MustParseURL("postgresql-2")},
 	}
 	anotherState.model = &mockModel{uuid: "uuid2", name: "another", owner: "fred"}
-	anotherState.usermodels = []crossmodel.UserModel{
+	anotherState.usermodels = []crossmodelcommon.UserModel{
 		&mockUserModel{model: anotherState.model},
 	}
 	anotherState.connStatus = &mockConnectionStatus{count: 15}
 
-	s.mockState.usermodels = []crossmodel.UserModel{
+	s.mockState.usermodels = []crossmodelcommon.UserModel{
 		&mockUserModel{model: s.mockState.model},
 		&mockUserModel{model: anotherState.model},
 	}
@@ -489,7 +358,7 @@ func (s *crossmodelSuite) TestFindMultiModel(c *gc.C) {
 	s.applicationOffers.CheckCallNames(c, listOffersBackendCall, listOffersBackendCall)
 }
 
-func (s *crossmodelSuite) TestFindError(c *gc.C) {
+func (s *remoteEndpointsSuite) TestFindError(c *gc.C) {
 	filter := params.OfferFilters{
 		Filters: []params.OfferFilter{
 			{
@@ -505,69 +374,6 @@ func (s *crossmodelSuite) TestFindError(c *gc.C) {
 	}
 
 	_, err := s.api.FindApplicationOffers(filter)
-	c.Assert(err, gc.ErrorMatches, fmt.Sprintf(".*%v.*", msg))
-	s.applicationOffers.CheckCallNames(c, listOffersBackendCall)
-}
-
-func (s *crossmodelSuite) assertList(c *gc.C, expectedErr error) {
-	s.setupOffers(c, "test")
-	filter := params.OfferFilters{
-		Filters: []params.OfferFilter{
-			{
-				OfferName:       "hosted-db2",
-				ApplicationName: "test",
-			},
-		},
-	}
-	found, err := s.api.ListApplicationOffers(filter)
-	if expectedErr != nil {
-		c.Assert(errors.Cause(err), gc.ErrorMatches, expectedErr.Error())
-		return
-	}
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(found, jc.DeepEquals, params.ListApplicationOffersResults{
-		[]params.ApplicationOfferDetails{
-			{
-				ApplicationOffer: params.ApplicationOffer{
-					ApplicationDescription: "description",
-					OfferName:              "hosted-db2",
-					OfferURL:               "fred/prod.hosted-db2",
-					Endpoints:              []params.RemoteEndpoint{{Name: "db"}},
-				},
-				CharmName:      "db2",
-				ConnectedCount: 5,
-			},
-		},
-	})
-	s.applicationOffers.CheckCallNames(c, listOffersBackendCall)
-}
-
-func (s *crossmodelSuite) TestList(c *gc.C) {
-	s.authorizer.Tag = names.NewUserTag("admin")
-	s.assertList(c, nil)
-}
-
-func (s *crossmodelSuite) TestListPermission(c *gc.C) {
-	s.assertList(c, common.ErrPerm)
-}
-
-func (s *crossmodelSuite) TestListError(c *gc.C) {
-	s.authorizer.Tag = names.NewUserTag("admin")
-	filter := params.OfferFilters{
-		Filters: []params.OfferFilter{
-			{
-				OfferName:       "hosted-db2",
-				ApplicationName: "test",
-			},
-		},
-	}
-	msg := "fail"
-
-	s.applicationOffers.listOffers = func(filters ...jujucrossmodel.ApplicationOfferFilter) ([]jujucrossmodel.ApplicationOffer, error) {
-		return nil, errors.New(msg)
-	}
-
-	_, err := s.api.ListApplicationOffers(filter)
 	c.Assert(err, gc.ErrorMatches, fmt.Sprintf(".*%v.*", msg))
 	s.applicationOffers.CheckCallNames(c, listOffersBackendCall)
 }
